refactor : console view 캡슐화
