--- src/main/java/com/zuniorteam/racingcar/view/GameInputView.java
@@ -1,7 +1,0 @@
-package com.zuniorteam.racingcar.view;
-
-import com.zuniorteam.racingcar.dto.GameInput;
-
-public interface GameInputView {
-    GameInput listen();
-}
