--- src/main/java/com/zuniorteam/racingcar/view/GameResultView.java
@@ -1,7 +1,0 @@
-package com.zuniorteam.racingcar.view;
-
-import com.zuniorteam.racingcar.dto.GameResult;
-
-public interface GameResultView {
-    void draw(GameResult gameResult);
-}
