+++ src/main/java/com/zuniorteam/racingcar/view/console/RacingGameBoard.java
@@ -1,11 +1,9 @@
-package com.zuniorteam.racingcar;
+package com.zuniorteam.racingcar.view.console;
 
 import com.zuniorteam.racingcar.core.RacingGame;
 import com.zuniorteam.racingcar.dto.GameInput;
 import com.zuniorteam.racingcar.dto.GameResult;
 import com.zuniorteam.racingcar.core.MovingStrategy;
-import com.zuniorteam.racingcar.view.GameInputView;
-import com.zuniorteam.racingcar.view.GameResultView;
 
 import java.util.Objects;
